podman: adapt run.sh
It doesn't work as a simple user, the build.sh and run.sh scripts need
to be run as root.

From a first test, podman is much faster than docker, to create the
container from the image, and then to run Pupil Capture. As a simple
user it's slower than docker, and Pupil Capture fails to run.