Update to Ubuntu 18.04 and Pupil 1.9
Ubuntu 18.10 has been released and is the latest version, but I prefer
to use the latest Ubuntu LTS version if it works for the Pupil, to avoid
the need to update every 6 months (and normally the LTS version is more
stable).

I've tested this commit, Pupil Capture runs fine. My host system is
Fedora 29.